fix tests, which seemingly broke on a circular import with new exports
